Fix #11 - Template applies -1 gain
By default the template didn't do anything. This can be confusing.

It is much more preferable for the template to do something simple
like apply a gain of -1. This way, when creating a new block, the
user can immediately take it for a spin.

Before this, running the new block without making any changes
would just crash the program.