CHAR_BUFF_SIZE macro and matrix error handle hack
The recent changes in the matrix lib have made rt_block_io
blow up.

First, any matrix error in a .conf would make the block exit. This
is dramatic during Xenomai Lab operation. The code was using the
unsafe version of matrix_new, so any malformed string would call
exit(1).
Changing this to the safe version fixes the mysterious crashes,
but introduces a error handling issue. Right now the errors are
handled localy, which is a maintenance nightmare.

There are two solutions to this problem.
	1-Introduce a rt_block_io interface to matrix operations.
	  This is only a subpar maintenance dream instead of a full
	  blown nightmare.
	2-Make the mtrx lib use the ERROR/DEBUG macros. This is
	  great for maintenance, but makes mtrx dependant on
	  rt_block_io.

A tough call.

Secondly, the matrix lib is storing doubles with %12.6f precision.
For a 4x4 matrix, this makes for a string that overloaded every
buffer in rt_block_io. And the error handling wasn't particularly
graceful. To make this maintainable, I've made every single
char buf[] in the code default to a CHAR_BUFF_SIZE of 300.
I believe the tradeoff makes sense. Yes, it uses more memory,
but now the code is under control.