Fix #35 - Push Matrix to 8x8 size
By changing the sizes of all queues and pipes from
hardcoded or (the now defunct) QUEUE_SIZE to sizeof(Matrix),
matrices of any size should be happily passed along between blocks.

HOWEVER, the buffer size CHAR_BUFFER_SIZE is still coded to 1024.
This is critical for settings operations. Don't have a solution
to make this dynamic. The reason is the buffer size is proportional
to the size of strings produced by matrix_string(). If %12.6f coding
is used or something like %1.3f ... the requirement changes. I don't
think it's possible to calculate.

The solution would be, of course, to make default buffer size
something like 2 or 4K but... that's a lot of memory.

Or, the solution could be not to use char buf[] at all, but use some
dynamicaly calculated string mallocated on the spot. Possible?
Hard to believe? Out of the realm of possibility?

On other hand, though, settings are made to be hand written. There's
no way in hell anyone will ever hand-write such a large Matrix. Hm.