Implement matrix_print_pretty()
Matrixes are a pain to read if they're all in a line. A typical
output of print_matrix is this:

	[1.000 0.000 0.000; 0.000 1.000 0.000; 0.000 0.000 1.000]

This function provides an interface to easily print the previous
example as
	   A
	|1 0 0|
	|0 1 0|
	|0 0 1|

Where "A" is the name of the string. The format of the numbers
(%1.3f, for example) is an input parameter.