Remove extra space in _pretty and defaul to %2.3f
This reverts changes from

commit 2e713d3a830354a65c4aeaeab093fbe3f32c1baa
    Print extra space before number in *print_pretty()

Which printed an extra space before positive numbers so that when
a negative number appeared, it would be lined with the positive

_1
-1

Where _ would be the extra space. The correct solution for this
problem is to use %2.0f, instead of %1.0f and adding spaces.
By using the propper format, any matrix can be properly lined up,
which is great.