Create global Makefile
Everything seems to be compiling. Makefile complexity facing the
user has been reduced to nada.

There is one remaining bug, however. If the build directories are
not already there, make won't create them and fail.

As it is now, this won't compile. I could add the empty
directories to the repository, but I'll try to fix the Makefile.