Fix error no missing build dir
Qt makes the dirs automatically, the (henceforth known as) block
Makefile does it by hand, but it works fine. It even rm -r the
thing on `make cleanall`.

I also rework the block Makefile a bit to make it more obivous it
builds two separate binaries, and that each of them may be built
in parallel. Should have some performance numbers soon.