Fix #38 - Refactor Makes (race conditions et al.)
This last commit fixes race condition and shares remaining .o's
and closes the makefile refactoring, ready to be merged into
mainline.

gain.o and gain_settings.o were being left in the block folder.
gain_settings.o is also needed by the GUI, so this object was
being compiled twice. Not anymore.

There was also a race condition between Qt's makefile and the
block one. If Qt created the build folders before the block,
the mkdir would fail. Now mkdir is called with a -p to avoid this.

--

Here's the final results between the old Makefiles and the one's
finished by this commit. Compile times of diagrams have been slashed
in half. Ironically, Xenomai Lab has a bug in the progress bar
tracking the build which hides this improvement. Here's the
data:

    blocks/tick$ time make
    real        0m6.591s
    user        0m5.276s
    sys 0m1.124s

    blocks/signal_generator$ time make
    real        0m6.359s
    user        0m5.164s
    sys 0m1.000s

    blocks/gain$ time make
    real        0m6.216s
    user        0m5.032s
    sys 0m1.020s

    blocks/display$ time make
    real        0m0.305s
    user        0m0.196s
    sys 0m0.080s

**TOTAL**: 19.47

    blocks/tick$ time make -j2
    real        0m4.563s
    user        0m5.288s
    sys 0m0.928s

    blocks/signal_generator$ time make -j2
    real        0m2.281s
    user        0m2.000s
    sys 0m0.408s

    blocks/gain$ time make -j2
    real        0m2.262s
    user        0m2.040s
    sys 0m0.340s

    blocks/display$ time make -j2
    real        0m0.345s
    user        0m0.380s
    sys 0m0.164s

**TOTAL**: 9.45