Fix #34 - No error on delete loop task
Issues a warning with DEBUG() instead.

Also stopped deleting the main task. It doesn't make sense because
the main task shadows the Linux process. If we try to kill it, it
instantly kills the running process. This is just silly. The main
task should _return_ from main, not kill itself like some norwegian
black metal singer.