Introduce Debug and respective Tick support
This is very unstable still. Code contains some warnings that
should be ironed out in the following commits. Some glitches
were seen in my preliminary tests and warrant further study.

Here's the gist of the new debug architecture.

Each block now has a "backdoor". Every time the main loop executes,
a progress report is sent to the "debug" queue. The debug block
periodically empties the queue of reports and dumps them to stdout
and (in future) to a text file. This text file will then plotted
visually so the user can interact with the data.

rt_block_io contains the library support. Several hooks were
introduced in key points throughtout the code flow to gather the
information for the report.
The report contains start and end times in ns from a monotinically
increasing time source. It also contains block name, input and
output matrices. This should be enough for now.
The report is serialized via a structure called a debugframe.
The "debug" queue supports some 10 of these debugframes.