Fix spurious crashes on diagram kill
This was a rather exotic bug. If the user clicked an element of
the GUI while "kill" was still running, the application would
crash.

The explication to why this happened is rather complicated.
First, a more accurate description of the bug.

When a user clicks one of the four action buttons on the toolbar,
MainWindow emits sendAction() which calls BlockDiagram's act()
slot.  When the action is finished, BlockDiagram emits finished()
which then calls MainWindow's backFromAction() slot [1].

If the user clicked any part of the main window while act() was
executing, when finished() was emited, backFromAction() would not
run. The application got stuck in a limbo. And this does not
happen with any other action, only "kill".

While I didn't understand precisely the root of the problem, I
was able to assert that this was due to the terminate() call on
the QProcess. This call would then generate a finished() signal
when the process exited. It was this signal mixed with the GUI
events that was throwing the application to a limbo.

The solution I found to this was to differ the terminate() call
untill after finished() was emited and backFromAction() executed.

Instead of calling terminate() directly, a killProcess() signal
is emited. This is a QueuedConnection, so it will only execute
when the event loop resumes.

At first killProcess() was calling the terminate() slot. This
works fine, but if the block was crashed for some unholy reason
disaster would ensue. Calling terminate() isn't enough, in
reality we want to call it and see if it worked, because if it
didn't we have to kill process. This is the only way to get the
system to a known state when the user clicks "kill". This
guarantees that all processes have exited.

To defer this sequence of steps, I encapsulated the steps in a
QThread. Then I had killProcess() call the start() slot of the
QThread. The thread calls terminate(), gives the process 2
seconds to comply. If it doesn't, it calls kill().

The architecture is considerably more complicated now, but the
bug is gone and the interface has gotten considerably more fluid.
Using "worker threads" I think is standard in GUIs precisely
because of this. Migrating further work from the main thread into
worker threads is an idea worth studying.

[1] This may seem rather convulated at first sight, i.e. why not
call the functions directly on the object and use signals/slots
if the operation is essentially sequential. While this doesn't
make as much sense in "kill", it makes more sense in the other
actions. During "make", BlockDiagram sends progress reports of
the build which are used to show a progress bar. This is
inherently asynchronous. During "exec", the diagram may execute
for hours. It makes more sense that the application moves to an
"execution" state and then move back to an "edit" state when
BlockDiagram signals it's finished, then to make it sequential.