Don't initialize queues with init value
This is the start of a considerable overall of the system.

Signals, or rather, queues have initialization values to solve the
problem of feedback loops. When a block tried to read a queue that
depended on its own output, it deadlocked. Example:

 ______________________
|      ___________     |
 -->  |           |    |
      |   BLOCK   |----
--->  |___________|

Initializiation solved this problem because that feedback signal
would have an initial bogus value. However, it also introduces a
major problem - the delay in signals is non-deterministic. Oon
complex diagrams, it's impossible to tell if the sample the block
reads is from the current cycle or from N cycles ago.

After much consideration, me and my collegues have reached the
conclusion that the best approach to solve the feedback problem
is not init values but delays blocks. Example:
       ___________
      |           |
 ---- |   Z^-1    |<---
|     |___________|    |
|      ___________     |
 -->  |           |    |
      |   BLOCK   |----
--->  |___________|

The major drawback is that the user can now design a diagram that
will not run. This is a major usability issue. This needs to be
very tightly controlled, both from the block library and XL itself.
With propper error support, we might be able to warn the user that
signal X has timedout and this is a sign of cyclical dependency.

Further study will have to be done on this. As an example to
illustrate the complexity, right now we can kill  'tick' and that
event alone will propagate down the diagram and make blocks exit
cleanly. There's a tight balance act between when a signal times
out and when its queue descriptor becomes invalid. Also consider
what happens when a block seg faults crashes. Every block down the
line will time out, even though it's not a cyclical dependecy
issue.

Like I said, further study is needed on this.