Force removal of Makefiles in settings_gui
If the build is interrupted for some reason, these Makefiles may be
left in a corrupt state. Yet, they won't update because there's no
reason to do so, dependency wise.

So, as a recap. When we run a `cleanall`, we remove the Makefiles,
generate them with QMake, run `clean` on them, and then remove them.