Improve handling of timeout in read_inputs()
Since queues don't have an initilization value anymore, it's now
possible that a user draws a diagram that just won't run. This
really ups the stakes in terms of error handling. This commit does a
general cleanup of the function, and uses a tick counter to test if
the timeout happened before the first sample or after it.

If it's before the first sample, this timeout was most likely caused
by cyclical dependency. It might also be that the parent block just
segfaulted before writing to the queue, but that's quite rare.

If it's after the first sample, it's clearly not from a cyclical
dependency.

Over the next commits a testsuite will be rolled in that will
explain and expose this behavior more specificaly.