Implement Z^-N block (delay)
This took a lot of work, especially on the testing side (which still
hasn't made it to the repository). Anyway, this commit adds the Z^-N
block and makes the necessary changes to rt_block_io.* to support
it.

The Z^-N block is, in itself, trivial. The only thing it does is
write an inital value N times to its output queues. To allow for
this simples implementation, the way queues are configured had to
change. Queues are now configured as FIFO, don't have initial
values, and allow are large enough for a decent number of samples to
be queued up.

Some notes on the implementation. First we rename MAX_MESSAGE_LENGTH
to MAX_QUEUE_LENGTH, which makes much more sense.

Each queue used to have a pool of memory of sizeof(Matrix) bytes,
and a maximum number of samples queued up MAX_QUEUE_LENGTH=1. They
are now configured as the debug queue, with a pool
MAX_QUEUE_LENGTH*sizeof(Matrix) and MAX_QUEUE_LENGTH possible
samples to queue up.

MAX_QUEUE_LENGTH has been set to 11, so we can set N up to 10. This
way Z can queue 10 initial values and the first result of whatever
comes before it.

The queue was set to FIFO operaton by changing the call to
rt_queue_write. Before it was Q_URGENT, which made the sample be
written on the top of the stack, but now it's Q_NORMAL, which
assures the FIFO behavior.